+++ lib/db.php
@@ -452,9 +452,7 @@ function glz_plugin_preferences($arr_preferences) {
 function glz_update_plugin_preferences($arr_preferences) {
   // die(dmp($arr_preferences));
   foreach ($arr_preferences as $preference => $value) {
-    // if there is a preference which doesn't exist, create it
-    if ( !isset($prefs[$preference]) )
-      set_pref($preference, addslashes(addslashes(trim($value))), "glz_custom_f", 10); // 10 so that it won't appear under TXP's prefs tab
+    set_pref($preference, addslashes(addslashes(trim($value))), "glz_custom_f", 10); // 10 so that it won't appear under TXP's prefs tab
   }
 }