+++ plugins/glz_custom_fields_code.php
@@ -1,0 +1,1 @@
+../glz_custom_fields_code.php