+++ releases/CHANGELOG.md
@@ -1,5 +1,13 @@
 What's new in v1.2.5
-==========================
+===
+ALL EXISTING PLUGIN PREFERENCES WILL BE OVERWRITTEN - RELATED TO PATH/URL CHANGES
+
 * no longer checking if jQuery is present - must be running TXP v4+
-* fixed installation on TXP 4.4.1
+* fixed installation for TXP 4.4.1
 * include_once fix for custom scripts
+* CSS & JS loads on the default page, right after user logs in GH-21
+* fixed editing of custom script names
+* CSS & JS code now in separate files
+* fixes duplicate excerpt bug when used together with MLP GH-23
+* disabling excerpts won't affect textarea custom fields, they are now
+  placed right after the article body GH-26